patch

+++ theme/views.py
@@ -38,12 +38,38 @@ def create_paid_member(request):
 # DELETE View
 @csrf_exempt
 @api_key_required
-def delete_paid_member(request, pk):
+def delete_paid_member(request, email):
     if request.method == "DELETE":
         try:
-            member = PaidMember.objects.get(pk=pk)
+            data = json.loads(request.body)
+            email = data.get('email')
+            if not email:
+                return JsonResponse({"error": "Email is required"}, status = 400)
+            member = PaidMember.objects.get(email=email)
             member.delete()
-            return JsonResponse({"message": f"Member {pk} deleted"}, status = 200)
+            return JsonResponse({"message": f"Member {email} deleted"}, status = 200)
         except PaidMember.DoesNotExist:
             return JsonResponse({"error": "Member not found"}, status = 404)
+        except json.JSONDecodeError:
+            return JsonResponse({"error": "Invalid JSON"}, status = 400)
+    return JsonResponse({"error": "Method not allowed"}, status = 405)
+
+@csrf_exempt
+@api_key_required
+def patch_paid_member(request, email):
+    if request.method == "PATCH":
+        try:
+            data = json.loads(request.body)
+            email = data.get('email')
+            canceled = data.get('canceled')
+            if not email and not canceled:
+                return JsonResponse({"error": "Email and cancel is required"}, status = 400)
+            member = PaidMember.objects.get(email=email)
+            member.canceled = canceled
+            member.save()
+            return JsonResponse({"message": f"Member {email} updated"}, status = 200)
+        except PaidMember.DoesNotExist:
+            return JsonResponse({"error": "Member not found"}, status = 404)
+        except json.JSONDecodeError:
+            return JsonResponse({"error": "Invalid JSON"}, status = 400)
     return JsonResponse({"error": "Method not allowed"}, status = 405)
